Added Reddit community link
Added a link for the newly created Reddit community, with the goal of having a place for users of Open VR Space Calibrator to work together and troubleshoot in a more public (indexed by search engines) and less linear forum than Discord. Hopefully this will increase the discoverability of mixed VR, help existing users of mixed VR troubleshoot issues by being able to Google them, as well as reduce some of the repetitive questions on the Discord.